+++ inst/doc/Logging.R
@@ -1,10 +1,10 @@
 ## ---- echo = FALSE, message = FALSE, warning = FALSE--------------------------
 library(ParallelLogger)
 knitr::opts_chunk$set(
-  cache = FALSE,
-  comment = "#>",
-  error = FALSE,
-  tidy = FALSE)
+cache = FALSE,
+comment = "#>",
+error = FALSE,
+tidy = FALSE)
 
 ## -----------------------------------------------------------------------------
 logger <- createLogger(name = "SIMPLE",
@@ -65,26 +65,39 @@ unlink(logFileName)
 #  addDefaultFileLogger(logFileName)
 
 ## ---- eval=FALSE--------------------------------------------------------------
-#  registerLogger(createLogger(name = "DEFAULT",
+#  registerLogger(createLogger(name = "DEFAULT_FILE_LOGGER",
 #                              threshold = "TRACE",
 #                              appenders = list(createFileAppender(layout = layoutParallel,
-#                                                                    fileName = logFileName))))
+#                                                                  fileName = logFileName))))
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  addDefaultErrorReportLogger()
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  fileName <- file.path(getwd(), "errorReportR.txt")
+#  
+#  registerLogger(createLogger(name = "DEFAULT_ERRORREPORT_LOGGER",
+#                              threshold = "FATAL",
+#                              appenders = list(createFileAppender(layout = layoutErrorReport,
+#                                                                  fileName = fileName,
+#                                                                  overwrite = TRUE,
+#                                                                  expirationTime = 60))))
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  mailSettings <- list(from = "someone@gmail.com",
-#                        to = c("someone_else@gmail.com"),
-#                        smtp = list(host.name = "smtp.gmail.com",
-#                                    port = 465,
-#                                    user.name = "someone@gmail.com",
-#                                    passwd = "super_secret!",
-#                                    ssl = TRUE),
-#                        authenticate = TRUE,
-#                        send = TRUE)
+#  to = c("someone_else@gmail.com"),
+#  smtp = list(host.name = "smtp.gmail.com",
+#  port = 465,
+#  user.name = "someone@gmail.com",
+#  passwd = "super_secret!",
+#  ssl = TRUE),
+#  authenticate = TRUE,
+#  send = TRUE)
 #  
 #  logger <- createLogger(name = "EMAIL",
-#                         threshold = "FATAL",
-#                         appenders = list(createEmailAppender(layout = layoutEmail,
-#                                                              mailSettings = mailSettings)))
+#  threshold = "FATAL",
+#  appenders = list(createEmailAppender(layout = layoutEmail,
+#  mailSettings = mailSettings)))
 #  registerLogger(logger)
 #  
 #  logFatal("No more data to process")
@@ -93,10 +106,10 @@ unlink(logFileName)
 #  addDefaultEmailLogger(mailSettings)
 
 ## ---- eval=FALSE--------------------------------------------------------------
-#   registerLogger(createLogger(name = "DEFAULT",
-#                               threshold = "FATAL",
-#                               appenders = list(createEmailAppender(layout = layoutEmail,
-#                                                                    mailSettings = mailSettings))))
+#  registerLogger(createLogger(name = "DEFAULT_EMAIL_LOGGER",
+#  threshold = "FATAL",
+#  appenders = list(createEmailAppender(layout = layoutEmail,
+#  mailSettings = mailSettings))))
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  clearLoggers()
@@ -130,11 +143,11 @@ addDefaultFileLogger(logFileName)
 cluster <- makeCluster(3)
 
 fun <- function(x) {
-  ParallelLogger::logInfo("The value of x is ", x)
-  # Do something
-  if (x == 6)
-    ParallelLogger::logDebug("X equals 6")
-  return(NULL)
+ParallelLogger::logInfo("The value of x is ", x)
+# Do something
+if (x == 6)
+ParallelLogger::logDebug("X equals 6")
+return(NULL)
 }
 
 dummy <- clusterApply(cluster, 1:10, fun, progressBar = FALSE)
@@ -142,6 +155,47 @@ dummy <- clusterApply(cluster, 1:10, fun, progressBar = FALSE)
 stopCluster(cluster)
 
 writeLines(readChar(logFileName, file.info(logFileName)$size))
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  subSubTask <- function() {
+#    addDefaultFileLogger("subSubTaskLog.txt")
+#    # do something
+#  }
+#  
+#  subTask <- function() {
+#    addDefaultFileLogger("subTaskLog.txt")
+#    # do something
+#    subSubTask()
+#  }
+#  
+#  mainTask <- function() {
+#    addDefaultFileLogger("mainTaskLog.txt")
+#    # do something
+#    subTask()
+#  }
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  subSubTask <- function() {
+#    # do something
+#  }
+#  
+#  subTask <- function() {
+#    # do something
+#    subSubTask()
+#  }
+#  
+#  mainTask <- function() {
+#    addDefaultFileLogger("log.txt")
+#    # do something
+#    subTask()
+#  }
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  mainTask <- function() {
+#    addDefaultFileLogger("log.txt")
+#    on.exit(unregisterLogger("DEFAULT_FILE_LOGGER"))
+#    # do something
+#  }
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  launchLogViewer(logFileName)